refactor(chat): neutralize run context service naming
