+++ turbo/apps/api/src/signals/services/agent-run-create.service.ts
@@ -258,7 +258,7 @@ import {
   type QueueFirstRunAssociation,
   type QueueFirstRunClaimResult,
   type QueueFirstRunSessionSnapshotState,
-} from "./zero-chat-queued-event.service";
+} from "./chat-queued-event.service";
 import { recordFirstAssistantEventEligibility } from "./zero-chat-first-assistant-event-metric.service";
 import { isWebChatTriggerSource } from "./chat-trigger-source.service";
 import { resolveVideoModelForRun } from "./video-model.service";
