+++ turbo/apps/api/src/signals/services/chat-queued-event.service.ts
