+++ turbo/apps/api/src/signals/services/web-chat-session-prompt.service.ts
@@ -22,7 +22,7 @@ import {
 import type { Db } from "../external/db";
 import { BEFORE_DISPATCH_CANCELLED_ERROR } from "./agent-run-create.service";
 import type { ChatThreadSessionResolutionAction } from "./chat-session-continuity.service";
-import { loadWebChatIncompleteContext } from "./zero-chat-incomplete-context.service";
+import { loadWebChatIncompleteContext } from "./chat-incomplete-context.service";
 import { visibleChatEventCondition } from "./zero-chat-event-shared.service";
 import {
   chatEventTextCondition,
