+++ turbo/apps/api/src/signals/services/zero-goal-queue-drain.service.ts
@@ -21,7 +21,7 @@ import {
   type PendingGoalQueueEvent,
 } from "./chat-goal-queue.service";
 import type { InternalRunCallbackKind } from "./internal-run-callback";
-import { resolveRunChatThreadModelContext } from "./zero-chat-run-event.service";
+import { resolveRunChatThreadModelContext } from "./chat-run-event.service";
 import { normalizeGoalObjectiveBrief } from "./goal-objective-brief-normalization.service";
 import type { ModelFirstPin } from "./zero-model-selection.service";
 import { createQueueFirstZeroRun$ } from "./zero-runs-create.service";
